separate buffer-format
